requirements.txt

pandas
openai
snowflake-connector-python
sqlite3
